+++ src/Container/ContainerSendersLocatorFactory.php
@@ -5,12 +5,7 @@ declare(strict_types=1);
 namespace Xtreamwayz\Expressive\Messenger\Container;
 
 use Psr\Container\ContainerInterface;
-use Symfony\Component\Messenger\Transport\Sender\SendersLocator;
-use function array_combine;
-use function array_keys;
-use function array_map;
-use function is_array;
-use function is_string;
+use Symfony\Component\Messenger\Transport\Sender\SendersLocatorInterface;
 
 final class ContainerSendersLocatorFactory
 {
@@ -22,36 +17,12 @@ final class ContainerSendersLocatorFactory
         $this->busName = $busName;
     }
 
-    public function __invoke(ContainerInterface $container) : SendersLocator
+    public function __invoke(ContainerInterface $container) : SendersLocatorInterface
     {
-        $config = $container->has('config') ? $container->get('config') : [];
-        $config = $config['messenger']['buses'][$this->busName] ?? [];
-
+        $config  = $container->has('config') ? $container->get('config') : [];
+        $config  = $config['messenger']['buses'][$this->busName] ?? [];
         $senders = $config['routes'] ?? [];
 
-        $sendersCallables = array_map(function ($senders) use ($container) {
-            $wrapper = function (string $senderId) use ($container) {
-                return function ($message) use ($container, $senderId) {
-                    $handler = $container->get($senderId);
-
-                    return $handler($message);
-                };
-            };
-
-            if (is_string($senders)) {
-                $senders = [$senders];
-            }
-
-            if (is_array($senders)) {
-                return array_map($wrapper, $senders);
-            }
-
-            throw new \InvalidArgumentException('Senders must be an array or string');
-        }, $senders);
-
-        return new SendersLocator(
-            array_combine(array_keys($senders), $sendersCallables),
-            []
-        );
+        return new ContainerSendersLocator($container, $senders, []);
     }
 }